+++ NEWS.md
@@ -27,6 +27,10 @@ Red Panda C++ Version 3.4
   - change: Show symbols defined in current file no matter it starts with underscore or not.
   - fix: Cpp Syntaxer is wrongly used to parse todo infos for everyfile.
   - fix: Wrongly indent when press tab in the middle of the line.
+  - Change: warn user when reload file with unsaved modifications.
+  - fix: Status of Toggle Block Comment not correctly updated.
+  - fix: Leading spaces is wrongly removed when paste contents in the middle of lines.
+  - change: Press Home/End to jump to Line begin/end when completion popup is shown.
   
 Red Panda C++ Version 3.3
   - enhancement: Auto indent contents in parenthesis.